PR feedback - add CORS layer - move logger to common crate - remove global log filters for nym-api and axum
