+++ common/http-api-common/Cargo.toml
@@ -13,8 +13,10 @@ license.workspace = true
 [dependencies]
 axum.workspace = true
 bytes = { workspace = true }
+colored.workspace = true
 mime = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
 serde_yaml = { workspace = true }
+tracing.workspace = true
 utoipa = { workspace = true }
